Fix: use npm^3 to install packages
This fixes the canary build on the Jenkins server by having it always use npm^3 internally when installing dependencies. (When running `npm test`, the local `npm` installation with the correct version will end up on the path.)

Eventually it would be a good idea to get this working with more recent npm versions, but this seems like a good first step in the meantime. The task should be significantly easier after we update configs to load plugins/shareable configs relative to themselves rather than ESLint.